Refactor provisoner to make it more modular.
