+++ src/Aspire.Hosting.Azure.Provisioning/AzureProvisionerExtensions.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Aspire.Hosting.Azure;
+using Aspire.Hosting.Azure.Provisioning;
 using Aspire.Hosting.Lifecycle;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Aspire.Hosting;
 
@@ -15,6 +17,18 @@ public static class AzureProvisionerExtensions
     public static IDistributedApplicationBuilder AddAzureProvisioning(this IDistributedApplicationBuilder builder)
     {
         builder.Services.AddLifecycleHook<AzureProvisioner>();
+        builder.AddAzureProvisioner<AzureKeyVaultResource, KeyVaultProvisoner>();
+        builder.AddAzureProvisioner<AzureStorageResource, StorageProvisioner>();
+        builder.AddAzureProvisioner<AzureServiceBusResource, ServiceBusProvisioner>();
+        return builder;
+    }
+
+    internal static IDistributedApplicationBuilder AddAzureProvisioner<TResource, TProvisioner>(this IDistributedApplicationBuilder builder)
+        where TResource : IAzureResource
+        where TProvisioner : AzureResourceProvisioner<TResource>
+    {
+        // This lets us avoid using open generics in the caller, we can use keyed lookup instead
+        builder.Services.AddKeyedSingleton<IAzuresourceProvisioner, TProvisioner>(typeof(TResource));
         return builder;
     }
 }
